requirements.txt
Django==3.1.3
python-dotenv==0.15.0
psycopg2==2.8.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
channels==3.0.2
channels-redis==3.2.0
asgiref==3.3.1